two commit

+++ src/com/ozkan/sumtwonumbers.java
@@ -2,9 +2,9 @@ package com.ozkan;
 
 public class sumtwonumbers{
         public static void main(String[] args) {
-            int num1 = 3;
-            int num2 = 4;
-            int result = num1 + num2;
+            double num1 = 3;
+            double num2 = 4;
+            double result = num1 + num2;
 
             System.out.println("Sum of two numbers : " + result);
         }
